@@ -3,29 +3,25 @@
 namespace Message\Envelope;
 
 use LogicException;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use SimpleBus\Serialization\Envelope\DefaultEnvelope;
 use SimpleBus\Serialization\Tests\Fixtures\DummyMessage;
 
 class DefaultEnvelopeTest extends TestCase
 {
-    /**
-     * @test
-     */
+    #[Test]
     public function itCreatesAnEnvelopeForAMessage(): void
     {
         $message = new DummyMessage();
         $type = get_class($message);
 
         $envelope = DefaultEnvelope::forMessage($message);
-        $this->assertInstanceOf(DefaultEnvelope::class, $envelope);
         $this->assertSame($message, $envelope->message());
         $this->assertSame($type, $envelope->messageType());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function itCreatesANewInstanceForADifferentMessage(): void
     {
         $message = new DummyMessage();
@@ -40,9 +36,7 @@ class DefaultEnvelopeTest extends TestCase
         $this->assertSame($type, $newEnvelope->messageType());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function itCreatesANewInstanceForASerializedVersionOfTheMessage(): void
     {
         $message = new DummyMessage();
@@ -58,9 +52,7 @@ class DefaultEnvelopeTest extends TestCase
         $this->assertSame($type, $newEnvelope->messageType());
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function itFailsWhenTheSerializedMessageIsUnavailable(): void
     {
         $message = new DummyMessage();
@@ -71,9 +63,7 @@ class DefaultEnvelopeTest extends TestCase
         $envelope->serializedMessage();
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function itFailsWhenTheMessageIsUnavailable(): void
     {
         $envelope = DefaultEnvelope::forSerializedMessage(
