@@ -3,6 +3,7 @@
 namespace SimpleBus\Serialization\Tests;
 
 use LogicException;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use SimpleBus\Serialization\Envelope\DefaultEnvelope;
 use SimpleBus\Serialization\NativeObjectSerializer;
@@ -11,24 +12,7 @@ use SimpleBus\Serialization\Tests\Fixtures\DummyMessage;
 
 class NativeObjectSerializerTest extends TestCase
 {
-    /**
-     * @test
-     */
-    public function itCanSerializeADefaultMessageEnvelopeWithASerializedMessage(): void
-    {
-        $envelope = DefaultEnvelope::forSerializedMessage(
-            DummyMessage::class,
-            'serialized message'
-        );
-        $serializer = new NativeObjectSerializer();
-
-        $serializedEnvelope = $serializedEnvelope = $serializer->serialize($envelope);
-        $this->assertIsString($serializedEnvelope);
-    }
-
-    /**
-     * @test
-     */
+    #[Test]
     public function itCanSerializeAndDeserializeADefaultMessageEnvelopeWithASerializedMessage(): void
     {
         $originalEnvelope = DefaultEnvelope::forSerializedMessage(
@@ -37,14 +21,12 @@ class NativeObjectSerializerTest extends TestCase
         );
         $serializer = new NativeObjectSerializer();
 
-        $serializedEnvelope = $serializedEnvelope = $serializer->serialize($originalEnvelope);
+        $serializedEnvelope = $serializer->serialize($originalEnvelope);
         $deserializedEnvelope = $serializer->deserialize($serializedEnvelope, get_class($originalEnvelope));
         $this->assertEquals($originalEnvelope, $deserializedEnvelope);
     }
 
-    /**
-     * @test
-     */
+    #[Test]
     public function itFailsWhenTheDeserializedObjectIsOfTheWrongType(): void
     {
         $expectedType = DummyMessage::class;
