@@ -2,6 +2,4 @@
 
 namespace SimpleBus\Serialization\Tests\Fixtures;
 
-class AnotherDummyMessage
-{
-}
+class AnotherDummyMessage {}
